+++ include/input/button.h
@@ -47,6 +47,6 @@ void button_tourney_lockout_enable(bool enable);
 void button_access_blocking(button_data_s *out, button_access_t type);
 void button_access_safe(button_data_s *out, button_access_t type);
 
-void button_task(uint32_t timestamp);
+void button_task(uint64_t timestamp);
 
 #endif
