+++ include/input/macros/macro_tourney.h
@@ -4,6 +4,6 @@
 #include "input_shared_types.h"
 #include <stdint.h>
 
-void macro_tourney(uint32_t timestamp, button_data_s *buttons);
+void macro_tourney(uint64_t timestamp, button_data_s *buttons);
 
 #endif
