+++ include/utilities/callback.h
@@ -5,8 +5,8 @@
 #include <stdbool.h>
 
 typedef void (*callback_t)(void);
-typedef void (*time_callback_t)(uint32_t);
+typedef void (*time_callback_t)(uint64_t);
 typedef bool(*hid_report_tunnel_cb)(uint8_t report_id, const void *report, uint16_t len);
-typedef void (*hid_task_tunnel_cb)(uint32_t timestamp, hid_report_tunnel_cb cb);
+typedef void (*hid_task_tunnel_cb)(uint64_t timestamp, hid_report_tunnel_cb cb);
 
 #endif
