+++ src/devices/rgb.c
@@ -135,7 +135,7 @@ void rgb_init(int mode, int brightness)
 
 // One tick of RGB logic
 // only performs actions if necessary
-void rgb_task(uint32_t timestamp)
+void rgb_task(uint64_t timestamp)
 {
     #if defined(HOJA_RGB_DRIVER) && (HOJA_RGB_DRIVER > 0)
     static interval_s interval = {0};
