+++ src/wired/wired.c
@@ -11,7 +11,7 @@
 
 time_callback_t _wired_task_cb = NULL;
 
-void wired_mode_task(uint32_t timestamp)
+void wired_mode_task(uint64_t timestamp)
 {
     if(_wired_task_cb)
         _wired_task_cb(timestamp);
